refactoring config classes
-removed unused old configs.
-updated imports on several classes
-moved foreign mod properties class
