+++ Treasure2-1.12.2/src/com/someguyssoftware/treasure2/command/SpawnWellStructureCommand.java
@@ -8,9 +8,7 @@ import java.util.Random;
 import com.someguyssoftware.gottschcore.positional.Coords;
 import com.someguyssoftware.gottschcore.positional.ICoords;
 import com.someguyssoftware.treasure2.Treasure;
-import com.someguyssoftware.treasure2.config.Configs;
 import com.someguyssoftware.treasure2.config.ModConfig;
-import com.someguyssoftware.treasure2.enums.Wells;
 import com.someguyssoftware.treasure2.generator.GeneratorData;
 import com.someguyssoftware.treasure2.generator.GeneratorResult;
 import com.someguyssoftware.treasure2.generator.well.WellGenerator;
