+++ Treasure2-1.12.2/src/com/someguyssoftware/treasure2/command/SpawnWitherTreeCommand.java
@@ -7,7 +7,6 @@ import java.util.Random;
 
 import com.someguyssoftware.gottschcore.positional.Coords;
 import com.someguyssoftware.treasure2.Treasure;
-import com.someguyssoftware.treasure2.config.Configs;
 import com.someguyssoftware.treasure2.config.ModConfig;
 import com.someguyssoftware.treasure2.generator.wither.WitherTreeGenerator;
